Removing broken verification inside survey.js rounter in the server side

+++ server/routes/survey.js
@@ -20,7 +20,10 @@ let Survey = require('../models/survey');
 // GET route for displaying all surveys
 router.get('/', function(req, res, next) {
   
-  this.blockIfUnauthorizedRequest();
+  if (!req.session.loggedin) {
+    res.status(401).send(`User not logged in`);
+    return;
+  }
 
   try {
     Survey.find()
@@ -41,7 +44,10 @@ router.get('/', function(req, res, next) {
 // GET route for survey results
 router.get('/results/:id', function(req, res, next) {
 
-  this.blockIfUnauthorizedRequest();
+  if (!req.session.loggedin) {
+    res.status(401).send(`User not logged in`);
+    return;
+  }
 
   try {
     Survey.findOne({ _id: req.params.id })
@@ -62,7 +68,10 @@ router.get('/results/:id', function(req, res, next) {
 // GET route for displaying a survey
 router.get('/update/:id', function(req, res, next) {
 
-  this.blockIfUnauthorizedRequest();
+  if (!req.session.loggedin) {
+    res.status(401).send(`User not logged in`);
+    return;
+  }
 
   try {
     Survey.findOne({ _id: req.params.id })
@@ -82,7 +91,10 @@ router.get('/update/:id', function(req, res, next) {
 // POST route for edit a survey
 router.post('/update', function(req, res, next) {
 
-  this.blockIfUnauthorizedRequest();
+  if (!req.session.loggedin) {
+    res.status(401).send(`User not logged in`);
+    return;
+  }
 
   let updatedSurvey = Survey({
     _id: req.body.id,
@@ -107,8 +119,11 @@ router.post('/update', function(req, res, next) {
 
 // POST route for create a survey
 router.post('/create', function(req, res, next) {
-
-  this.blockIfUnauthorizedRequest();
+  
+  if (!req.session.loggedin) {
+    res.status(401).send(`User not logged in`);
+    return;
+  }
 
   let questions = req.body.questions.map((q) => ({
     questionType: q.questionType,
@@ -130,7 +145,10 @@ router.post('/create', function(req, res, next) {
 // GET route to delete a survey
 router.get('/delete/:id', function(req, res, next) {
 
-  this.blockIfUnauthorizedRequest();
+  if (!req.session.loggedin) {
+    res.status(401).send(`User not logged in`);
+    return;
+  }
 
   Survey.deleteOne({ _id: req.params.id })
           .then((data) => {
@@ -142,11 +160,5 @@ router.get('/delete/:id', function(req, res, next) {
             }
           });
 });
-
-function blockIfUnauthorizedRequest() {
-  if (!req.session.loggedin) {
-    res.status(401).send(`User not logged in`);
-  }
-}
 
 module.exports = router;
